dom/tests/mochitest/bugs/test_bug918719.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=918719
-->
<head>
  <title>Test for Bug 918719</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=918719">Mozilla Bug 918719</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

function runTest() {
  return new Promise(function(resolve) {
    let loadingEventCount = 0,
        xhr = new XMLHttpRequest();
    xhr.open("GET", "bug918719.sjs");
    xhr.onreadystatechange = function() {
      if (xhr.readyState == 3) {
        loadingEventCount++;
      } else if (xhr.readyState == 4) {
        resolve(loadingEventCount);
      }
    }
    xhr.send();
  });
}

function prefChangePromise(args) {
  return new Promise(function(resolve) {
    SpecialPowers.pushPrefEnv(args, resolve);
  });
}

runTest().then(function(count) {
  ok(count === 1, "Only one loading readystatechange event should have been fired with the pref off.");
}).then(function() {
  return prefChangePromise({"set": [["dom.fire_extra_xhr_loading_readystatechanges", true]]});
}).then(function() {
  return runTest();
}).then(function(count) {
  // Note this case is tricky to test, as there is no hard guarantee that the
  // server will send the data in more than one chunk, even with the trick
  // we're using in the sjs. As such this case *may* intermittently fail.
  ok(count > 1, "Multiple loading readystatechange events should have been fired with the pref on.");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
